Tests/LibWeb/Text/expected/HTML/document-close-complete-readystatechange-replaces-parser.txt

Replaced parser during complete readystatechange event
Replacement ready state before close: loading
Iframe load event fired with replacement ready state complete
